+++ src/CodeGenHero.DataService/Models/WebApiDataServiceBase.cs
@@ -5,6 +5,7 @@ using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using cghConstants = CodeGenHero.DataService.Constants;
 using CodeGenHero.Logging;
+using Newtonsoft.Json;
 
 namespace CodeGenHero.DataService
 {
@@ -102,8 +103,10 @@ namespace CodeGenHero.DataService
 
 		public virtual HttpClient GetClient(AuthenticationHeaderValue authorization, int requestedVersion, string connectionIdentifier)
 		{
-			HttpClient client = new HttpClient();
-			client.BaseAddress = new Uri(ExecutionContext.BaseWebApiUrl);
+			HttpClient client = new HttpClient
+			{
+				BaseAddress = new Uri(ExecutionContext.BaseWebApiUrl)
+			};
 			client.DefaultRequestHeaders.Accept.Clear();
 			client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
 
@@ -151,24 +154,32 @@ namespace CodeGenHero.DataService
 		}
 
 		protected virtual async Task<List<T>> GetAllPageDataResultsAsync<T>(IPageDataRequest pageDataRequest, Func<IPageDataRequest,
-			Task<IPageDataT<List<T>>>> getMethodToRun)
+			Task<IHttpCallResultCGHT<IPageDataT<IList<T>>>>> getMethodToRun, bool throwExceptionOnFailureStatusCode = false)
 		{
 			List<T> retVal = new List<T>();
-			IPageDataT<List<T>> results = null;
+			IHttpCallResultCGHT<IPageDataT<IList<T>>> response = null;
 			IPageDataRequest currentPageDataRequest = new PageDataRequest(pageDataRequest.FilterCriteria, pageDataRequest.Sort, pageDataRequest.Page, pageDataRequest.PageSize);
 
-			while (results == null
-				|| (results.IsSuccessStatusCode == true && currentPageDataRequest.Page <= results.TotalPages))
+			while (response == null
+				|| (response.IsSuccessStatusCode == true && currentPageDataRequest.Page <= response.Data.TotalPages))
 			{
-				results = await getMethodToRun(currentPageDataRequest);
-				if (results.IsSuccessStatusCode)
+				response = await getMethodToRun(currentPageDataRequest);
+				if (response.IsSuccessStatusCode && response.Data != null)
 				{
-					retVal.AddRange(results.Data);
+					retVal.AddRange(response.Data.Data);
 				}
 				else
 				{
-					Log.Error($"Failure detected during data retrieval with currentPage = {currentPageDataRequest.Page} pageSize = {currentPageDataRequest.PageSize}.",
-						LogMessageType.Instance.Exception_WebApi);
+					string serializedCurrentPageDataRequest = JsonConvert.SerializeObject(currentPageDataRequest);
+					string msg = $"{nameof(GetAllPageDataResultsAsync)} call resulted in error - status code: {response?.StatusCode}; reason: {response?.ReasonPhrase}. CurrentPageDataRequest: {serializedCurrentPageDataRequest}";
+
+					Log.Error(message: msg, logMessageType: LogMessageType.Instance.Exception_WebApiClient, ex: response?.Exception,
+						httpResponseStatusCode: (int)response?.StatusCode, url: null);
+
+					if (throwExceptionOnFailureStatusCode == true)
+					{
+						throw new ApplicationException(msg, response?.Exception);
+					}
 				}
 
 				currentPageDataRequest.Page += 1;