+++ src/CodeGenHero.Repository/Interfaces/ICRUDOperation.cs
@@ -1,17 +1,18 @@
 ﻿using System.Linq;
+using System.Threading.Tasks;
 
 namespace CodeGenHero.Repository
 {
 	public interface ICRUDOperation<EntityType> where EntityType : class
 	{
-		RepositoryActionResult<EntityType> Delete(EntityType item);
+		Task<RepositoryActionResult<EntityType>> DeleteAsync(EntityType item);
 
-		EntityType GetFirstOrDefault(EntityType item);
+		Task<EntityType> GetFirstOrDefaultAsync(EntityType item);
 
 		IQueryable<TEntity> GetQueryable<TEntity>() where TEntity : class;
 
-		RepositoryActionResult<EntityType> Insert(EntityType item);
+		Task<RepositoryActionResult<EntityType>> InsertAsync(EntityType item);
 
-		RepositoryActionResult<EntityType> Update(EntityType item);
+		Task<RepositoryActionResult<EntityType>> UpdateAsync(EntityType item);
 	}
 }